# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ConnectionMode = typeof ConnectionMode[keyof typeof ConnectionMode];


export const ConnectionMode = {
  none: 'none',
  tunnel: 'tunnel',
  ngrok: 'ngrok',
} as const;

export type ProcessState = typeof ProcessState[keyof typeof ProcessState];


export const ProcessState = {
  stopped: 'stopped',
  starting: 'starting',
  running: 'running',
  error: 'error',
} as const;

export interface Server {
  id: number;
  name: string;
  command: string;
  args: string[];
  connectionMode: ConnectionMode;
  state: ProcessState;
  /** @nullable */
  publicUrl?: string | null;
  /** @nullable */
  lastError?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ServerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  command: string;
  args?: string[];
  connectionMode?: ConnectionMode;
}

export interface ServerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  command?: string;
  args?: string[];
  connectionMode?: ConnectionMode;
}

export type EnvVarScope = typeof EnvVarScope[keyof typeof EnvVarScope];


export const EnvVarScope = {
  global: 'global',
  server: 'server',
} as const;

export interface EnvVar {
  id: number;
  /** @nullable */
  serverId?: number | null;
  scope: EnvVarScope;
  key: string;
  value: string;
}

export interface EnvVarInput {
  /** @nullable */
  serverId?: number | null;
  /** @minLength 1 */
  key: string;
  value: string;
}

export interface EnvVarUpdate {
  /** @minLength 1 */
  key?: string;
  value?: string;
}

export interface TunnelConfig {
  serverId: number;
  tunnelId: string;
  hasApiKey: boolean;
  /** @nullable */
  uiUrl?: string | null;
}

export interface TunnelConfigInput {
  /** @minLength 1 */
  tunnelId: string;
  apiKey?: string;
}

export interface NgrokConfig {
  serverId: number;
  hasAuthToken: boolean;
  bearerToken: string;
  /** @nullable */
  domain?: string | null;
  /** @nullable */
  publicUrl?: string | null;
}

export interface NgrokConfigInput {
  authToken?: string;
  domain?: string;
  rotateBearerToken?: boolean;
}

export interface RuntimeStatus {
  serverId: number;
  state: ProcessState;
  connectionMode: ConnectionMode;
  /** @nullable */
  pid?: number | null;
  /** @nullable */
  publicUrl?: string | null;
  /** @nullable */
  connectorUrl?: string | null;
  /** @nullable */
  bearerToken?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  lastError?: string | null;
}

export type LogLineStream = typeof LogLineStream[keyof typeof LogLineStream];


export const LogLineStream = {
  stdout: 'stdout',
  stderr: 'stderr',
  system: 'system',
} as const;

export interface LogLine {
  timestamp: string;
  stream: LogLineStream;
  message: string;
}

export interface LogBundle {
  serverId: number;
  lines: LogLine[];
}

export interface Summary {
  totalServers: number;
  runningServers: number;
  stoppedServers: number;
  erroredServers: number;
  tunnelServers: number;
  ngrokServers: number;
  globalEnvVarCount: number;
}

export interface EnvironmentInfo {
  platform: string;
  tunnelClientInstalled: boolean;
  /** @nullable */
  tunnelClientVersion?: string | null;
  ngrokInstalled?: boolean;
  boundToLocalhost: boolean;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

export type ListEnvVarsParams = {
serverId?: number;
};