# lib/api-zod/src/generated/types/envVarInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnvVarInput {
  /** @nullable */
  serverId?: number | null;
  /** @minLength 1 */
  key: string;
  value: string;
}