# lib/api-zod/src/generated/types/logLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogLineStream } from './logLineStream';

export interface LogLine {
  timestamp: string;
  stream: LogLineStream;
  message: string;
}