# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionMode } from './connectionMode';
import type { ProcessState } from './processState';

export interface Server {
  id: number;
  name: string;
  command: string;
  args: string[];
  connectionMode: ConnectionMode;
  state: ProcessState;
  /** @nullable */
  publicUrl?: string | null;
  /** @nullable */
  lastError?: string | null;
  createdAt: string;
  updatedAt: string;
}