# lib/api-zod/src/generated/types/serverUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionMode } from './connectionMode';

export interface ServerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  command?: string;
  args?: string[];
  connectionMode?: ConnectionMode;
}